fix: cross-compiler and static-analysis failures from the develop merge
MSVC was the only compiler this branch had ever been built with locally,
so the merge into develop (which triggers GCC/Clang/CodeQL/clang-tidy legs
MSVC-only work never exercised) surfaced four real issues, each verified
against real toolchains (WSL2 gcc-15/clang-21) before pushing:

- transient_prenoise.cpp: CrossFade<N>'s int template parameter converting
  to std::array's size_t one tripped -Wsign-conversion under GCC (not
  caught by MSVC or clang). Made N a std::size_t directly.
- test_transient_prenoise.cpp: two float->double promotions on a
  std::vector<float> element read tripped -Wdouble-promotion under Clang.
  Made the promotion explicit.
- docs/index.md: a relative link to ../README.md broke `mkdocs build
  --strict` (README.md sits outside docs_dir, so mkdocs won't resolve
  it) - a pre-existing bug from earlier in this branch's history, not
  something the merge introduced. Switched to the GitHub blob URL every
  other README cross-reference in this docs tree already uses.
- main.cpp: three call sites dereferenced std::optional (via the
  std::expected<std::optional<T>, DecodeError> decode_access_unit now
  returns) through (*decoded)->member after a `continue`-guarded
  has_value() check clang-tidy's bugprone-unchecked-optional-access
  couldn't connect to the access - correct at runtime but the loop's
  own late-bound optional-of-optional pattern this API required.
  Bound the checked value to a local reference once per site instead,
  which both satisfies the analyzer directly and reads better than the
  repeated (*decoded)-> chain.

Verified against real gcc-15 and clang-21 builds (WSL2), not just MSVC:
full rebuild + 349/349 tests pass on both, plus the exact clang-tidy
invocation CI uses (run-clang-tidy-21 -warnings-as-errors='*' over
src/(lib|cli|matroska)) clean across all 79 matched files.

+++ tests/test_transient_prenoise.cpp
@@ -136,7 +136,7 @@ TEST_CASE("apply_transient_prenoise reduces energy where a transient's pre-noise
     double noisy_energy = 0.0;
     for (int samp = ac3::kTransientPrenoiseTC1; samp < tot_corr_len - ac3::kTransientPrenoiseTC2;
         ++samp) {
-        const double v = pcm[static_cast<std::size_t>(start_samp + samp)];
+        const double v = static_cast<double>(pcm[static_cast<std::size_t>(start_samp + samp)]);
         noisy_energy += v * v;
     }
 
@@ -145,7 +145,7 @@ TEST_CASE("apply_transient_prenoise reduces energy where a transient's pre-noise
     double corrected_energy = 0.0;
     for (int samp = ac3::kTransientPrenoiseTC1; samp < tot_corr_len - ac3::kTransientPrenoiseTC2;
         ++samp) {
-        const double v = pcm[static_cast<std::size_t>(start_samp + samp)];
+        const double v = static_cast<double>(pcm[static_cast<std::size_t>(start_samp + samp)]);
         corrected_energy += v * v;
     }
     CHECK(corrected_energy < noisy_energy * 0.01);
